requirements.txt
mysql-connector-python
vt-py